Move shakeListener to DI
